+++ src/main/java/com/anipick/backend/anime/service/dto/ComingSoonRequestDto.java
@@ -10,11 +10,11 @@ import java.time.LocalDate;
 public class ComingSoonRequestDto {
     private Long lastId;
     private String lastValue;
-    private int size;
-    private int includeAdult;
+    private Long size;
+    private Long includeAdult;
     private String orderByQuery;
 
-    public static ComingSoonRequestDto of(Long lastId, String lastValue, int size, int includeAdult, String orderByQuery) {
+    public static ComingSoonRequestDto of(Long lastId, String lastValue, Long size, Long includeAdult, String orderByQuery) {
         return new ComingSoonRequestDto(lastId, lastValue, size, includeAdult, orderByQuery);
     }
 }
